Fix elIsOnScreen
The function was checking if the el's Y position is greater than its
height. Its Y position is greater than its height when the space
between the top of the viewport and the top of its outter-most
box is greater than its height. This is because the origin of the
Y position of an element is its top most box.